+++ src/public/src/js/yap-core.js
@@ -169,9 +169,10 @@ function getMetricsData()
 {
     $("#metrics").slideUp(function () {
         $.getJSON("../api/v1/reports/metrics?service_body_id=" + $("#service_body_id").val() + getDateRanges() + "&recurse=" + recurseReports(), function (data) {
-            var actions = ['Volunteer (CALL)', 'Meetings (CALL)', 'Just For Today (CALL)', 'Volunteer (SMS)', 'Meetings (SMS)', 'Just For Today (SMS)'];
-            var actions_plots = [1, 2, 3, 19, 20, 21];
-            var plots = {"1": [], "2": [], "3": [], "19": [], "20": [], "21": []};
+            var actions = ['Volunteer (CALL)', 'Meetings (CALL)', 'JFT (CALL)', 'Volunteer (SMS)', 'Meetings (SMS)', 'JFT (SMS)', 'SPAD', 'SPAD (SMS)'];
+            var actions_plots = [1, 2, 3, 19, 20, 21, 23, 24];
+            var plots = {"1": [], "2": [], "3": [], "19": [], "20": [], "21": [], "23": [], "24": []};
+            var colors = ['#FF6600', '#87B63A', 'indigo', '#FF6E9B', '#446E9B', 'black', 'purple', 'brown'];
             for (let item of data['metrics']) {
                 plots[JSON.parse(item['data'])['searchType']].push({
                     'x': item['timestamp'],
@@ -210,7 +211,6 @@ function getMetricsData()
             $("#summary-missedvolunteer-calls").html(missedCalls + " ("+ missedCallsPct + "%)");
 
             var datasets = [];
-            var colors = ['#FF6600', '#87B63A', 'indigo', '#FF6E9B', '#446E9B', 'black'];
             for (var a = 0; a < actions.length; a++) {
                 var xAgg = [];
                 var yAgg = [];